requirements.txt
# Python 3.10+ recommandé
torch==2.0.1
Pillow==9.5.0
rembg==2.0.50

